# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retail POS SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  businessName: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  manager: 'manager',
  cashier: 'cashier',
} as const;

export interface User {
  id: number;
  tenantId: number;
  email: string;
  name: string;
  role: UserRole;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Tenant {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  logoUrl?: string | null;
  currency?: string;
  timezone?: string;
  createdAt: string;
}

export interface TenantUpdate {
  name?: string;
  /** @nullable */
  logoUrl?: string | null;
  currency?: string;
  timezone?: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  manager: 'manager',
  cashier: 'cashier',
} as const;

export interface UserInput {
  email: string;
  name: string;
  role: UserInputRole;
  /** @minLength 6 */
  password: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  manager: 'manager',
  cashier: 'cashier',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
}

export interface Category {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  description?: string;
  color?: string;
}

export interface CategoryUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Product {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sku?: string | null;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  sku?: string;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sku?: string | null;
  price?: number;
  stock?: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  tenantId: number;
  orderNumber: string;
  status: OrderStatus;
  subtotal: number;
  tax: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cashierId?: number | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  items: OrderItemInput[];
  paymentMethod?: string;
  customerName?: string;
  notes?: string;
  tax?: number;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  paymentMethod?: string;
  notes?: string;
}

export interface DashboardSummary {
  todayRevenue: number;
  todayOrders: number;
  totalProducts: number;
  totalCustomers: number;
  revenueChange: number;
  ordersChange: number;
}

export interface SalesByDay {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  totalRevenue: number;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
limit?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;