# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Retail POS SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new tenant and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a product category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, completed, cancelled, refunded]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegisterInput:
      type: object
      required: [email, password, name, businessName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        businessName:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, tenantId, email, name, role, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, cashier]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    Product:
      type: object
      required: [id, tenantId, name, price, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
          format: float
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
          format: float
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, subtotal]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        subtotal:
          type: number
          format: float
    Order:
      type: object
      required: [id, tenantId, orderNumber, status, subtotal, tax, total, items, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled, refunded]
        subtotal:
          type: number
          format: float
        tax:
          type: number
          format: float
        total:
          type: number
          format: float
        paymentMethod:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cashierId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    OrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        paymentMethod:
          type: string
        customerName:
          type: string
        notes:
          type: string
        tax:
          type: number
          format: float
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, cancelled, refunded]
        paymentMethod:
          type: string
        notes:
          type: string
...