# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retail POS SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  tenantId: number;
  orderNumber: string;
  status: OrderStatus;
  subtotal: number;
  tax: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cashierId?: number | null;
  items: OrderItem[];
  createdAt: Date;
}