# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retail POS SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sku?: string | null;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: Date;
}