# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retail POS SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  sku?: string;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  imageUrl?: string;
}