# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retail POS SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  tenantId: number;
  email: string;
  name: string;
  role: UserRole;
  isActive?: boolean;
  createdAt: Date;
}